Sessions in directories now, named like Atl_ITCZ
One example Atlantic ITCZ is now included, along with a rename_me blank
notebook suggesting the user to do their own.

It was a pain to make ClickHistDo write Session as directories instead
of longer filenames, but this will be much more useful.

Tested Dec 2016, Hamburg, flying out tomorrow. End of a week of work.